pyproject.toml
[tool.poetry]
name = "goblin_fetcher"
version = "0.1.0"
description = "Data fetcher for externally specified GOBLIN database"
authors = ["Colm Duffy"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "2.1.4"
numpy = "^1.25.0"
sqlalchemy-utils = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
